[G2J-007] Moved hook parsing to own pkg
